Change pairing_cb to assume temporary pairing by default
When pairing takes place, the pairing_cb.is_temp flag indicates whether
a pairing is temporary or permanent. Link keys are not stored for
temporary pairings. Since this is a "positive" flag, resetting the
pairing_cb control block (ex. memset to 0), it will assume persistent
pairing by default. Under certain circumstances, this can lead to a link
key being stored for temporarily secured connection.

This patch reverses the flag to be a "negative" flag. Renamed to
"persistent_bond", the default 0 meaning is now used to indicate a
temporary bond. If the flag is not properly set now, it will default to a
temporary bond and will not save the link key erronously.

This fix is for CVE-2014-7914

Bug: 18345373
Change-Id: I6e821595877ff30d64dc6e33602ac049cab3dd1e

Conflicts:
	btif/src/btif_dm.c